hw4/app/requirements.txt
streamlit==1.24.0
psycopg2-binary==2.9.6